home_rent/requirements.txt
Django==1.11
djangorestframework==3.7.4
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
pytz==2017.3
